feat: merge the two tracked changes demos
