+++ next.config.ts
@@ -1,7 +1,15 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  async redirects() {
+    return [
+      {
+        source: "/server-ai-tracked-changes-comments",
+        destination: "/server-ai-tracked-changes",
+        permanent: true,
+      },
+    ];
+  },
 };
 
 export default nextConfig;
