--- src/app/api/server-ai-tracked-changes-comments/route.ts
@@ -1,110 +1,0 @@
-import { devToolsMiddleware } from "@ai-sdk/devtools";
-import {
-  createAgentUIStreamResponse,
-  gateway,
-  ToolLoopAgent,
-  tool,
-  type UIMessage,
-  wrapLanguageModel,
-} from "ai";
-import z from "zod";
-import { getIp, rateLimit } from "@/lib/rate-limit";
-import { executeTool } from "@/lib/server-ai-toolkit/execute-tool";
-import { getTools } from "@/lib/server-ai-toolkit/get-tools";
-
-export async function POST(req: Request) {
-  if (process.env.UPSTASH_REDIS_REST_URL) {
-    const ip = await getIp();
-    const isAllowed = await rateLimit(ip);
-
-    if (!isAllowed) {
-      return new Response("Rate limit exceeded. Please try again later.", {
-        status: 429,
-        headers: {
-          "Content-Type": "text/plain",
-        },
-      });
-    }
-  }
-
-  const {
-    messages,
-    editorContext,
-    documentId,
-  }: {
-    messages: UIMessage[];
-    editorContext: unknown;
-    documentId: string;
-  } = await req.json();
-  const toolsResponse = await getTools({
-    editorContext,
-    operationMeta:
-      "Brief justification explaining why this change improves the document.",
-  });
-
-  const tools = Object.fromEntries(
-    toolsResponse.tools.map((toolDef) => [
-      toolDef.name,
-      tool({
-        description: toolDef.description,
-        inputSchema: z.fromJSONSchema(toolDef.inputSchema),
-        execute: async (input) => {
-          try {
-            const result = await executeTool(
-              toolDef.name,
-              input,
-              null,
-              editorContext,
-              {
-                documentId,
-                userId: "ai-assistant",
-                reviewOptions: {
-                  mode: "trackedChanges",
-                },
-                commentsOptions: {
-                  threadData: { userName: "Tiptap AI" },
-                  commentData: { userName: "Tiptap AI" },
-                },
-              },
-            );
-
-            return result.output;
-          } catch (error) {
-            console.error(`Failed to execute tool ${toolDef.name}:`, error);
-            return {
-              error: error instanceof Error ? error.message : "Unknown error",
-            };
-          }
-        },
-      }),
-    ]),
-  );
-
-  const model = wrapLanguageModel({
-    model: gateway("openai/gpt-5.4-mini"),
-    middleware:
-      process.env.NODE_ENV === "production" ? [] : devToolsMiddleware(),
-  });
-
-  const agent = new ToolLoopAgent({
-    model,
-    instructions: `You are an assistant that can edit rich text documents with tracked changes and linked Tiptap comments.
-In your messages to the user, be concise and to the point. However, the content of the document you generate does not need to be concise and to the point, instead, it should follow the user's request as closely as possible.
-Before calling any tools, summarize what you're going to do in one short sentence.
-Rule: Use tiptapRead before tiptapEdit.
-Rule: Keep user-facing responses to a single short sentence before tool calls and a single short sentence after completion.
-Rule: In your messages to the user, do not give any details of the tool calls.
-Rule: In your messages to the user, do not give any details of the document content, the individual edits, or the justifications for those edits.
-Rule: In your messages to the user, never mention hashes, tool internals, raw document JSON, or where to find the comments in the UI.
-Rule: For every tiptapEdit operation, always provide a brief justification in the meta field explaining why the change improves the document.
-Rule: Put justifications only in the meta field so they become linked Tiptap Comments. Do not repeat those justifications in assistant messages.
-
-${toolsResponse.prompt}`,
-    tools,
-  });
-
-  return createAgentUIStreamResponse({
-    agent,
-    uiMessages: messages,
-  });
-}
