+++ src/app/page.tsx
@@ -186,14 +186,9 @@ const CATEGORIES = [
           },
           {
             title: "Server Tracked Changes",
-            description: "Review server-side AI edits as tracked changes",
-            href: "/server-ai-tracked-changes",
-          },
-          {
-            title: "Server Tracked + Comments",
             description:
               "Tracked changes with server-side comments explaining each edit",
-            href: "/server-ai-tracked-changes-comments",
+            href: "/server-ai-tracked-changes",
           },
         ],
       },
